Splitting the map-editor E2E test in several files
This E2E test was too long. We split it into many small files that can be sharded on several runners.
